db: hold the ledger in base units
The ledger records base units of an agent's stake token, not chips. It is the
money record and it has to reconcile one for one with a vault, and a vault
holds tokens.

src/chips.ts is the whole conversion, and says why no chip balance is stored
anywhere: an agent's chips are computed from its base units when they are
needed, so at a season boundary the same tokens simply count as a different
number of chips and there is no conversion event to stand in front of. The
design that stores chips and converts them to tokens is the one that can be
farmed - deposit while chips are cheap in tokens, withdraw while they are dear.
Rounding exists only in the view: an amount that began as chips is an exact
multiple of the rate, and what can have a remainder is a balance, because a
deposit or a withdrawal is any number of base units the owner chose.

The columns widen to bigint with no backfill. Six decimals put a 900-chip
balance at 9e8 and a 32-bit column tops out near 2,147 chips - but the frozen
seed program's rate is 1, so every row already written is a correct base-unit
value and only the type had to change.

Two things the test found that widening the columns alone did not fix:

balanceOf and every other money sum still cast to ::int, so a wide column was
read through a narrow cast and overflowed at the same 2,147 chips. The sums
over match nets and stakes keep ::int, because those stay in chips.

node-postgres reads an int8 as a string, while PGlite reads it as a number.
Through a typed column drizzle hides the difference; through a raw
sum(...)::bigint it does not - so a balance would have come back as
"900000000" in production and every test, all of which run on PGlite, would
have passed. The two are made to agree in one place, with anything too large to
be exact throwing rather than quietly losing its last digits. The pg path could
not be exercised here - no docker, and the Postgres that is listening is not
this project's - so it is verified at the parser instead.

+++ src/db/house.ts
@@ -27,7 +27,7 @@ export async function pickHousePair(
 ): Promise<[string, string] | null> {
   // Joined rather than correlated, as in playableBands.
   const balances = db
-    .select({ agentId: ledger.agentId, balance: sql<number>`sum(${ledger.amount})::int`.as("balance") })
+    .select({ agentId: ledger.agentId, balance: sql<number>`sum(${ledger.amount})::bigint`.as("balance") })
     .from(ledger)
     .groupBy(ledger.agentId)
     .as("balances");
